Factor markPlaying method to avoid duplicate code
